feat: history generator
This commit introduces a new class (DatagenHistoryGenerator)
which creates a week's worth of historical events.

The role of this class is to orchestrate the different
generators, using each of them to create a history for different
event types, and then combine them into a single sorted list of
SourceRecord objects ready for delivering to Kafka Connect.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>