feat: generate historical events on startup (#7)
Typical Event Processing demos often require several days worth of events
on Kafka topics, as these allow us to demonstrate how we can identify 
infrequent logical events that span multiple topics. 

This means it is necessary to run the datagen connector for several days
before some of these demos can be given, to give the datagen connector
time to incrementally build up a history of events. 

This pull request provides an alternate strategy - when starting up for the
first time, the connector can now optionally generate a week's worth of
historical events, with timestamps set to make events appear consistent 
with what the connector would have generated if it had been running for a 
week.

This new behavior is optional, and disabled by default. 

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>
Co-authored-by: Thajinu Shehin <Thajinu.Shehin@ibm.com>