Online orders cities (#5)
We want to display the shipping addresses of online orders on Google 
Maps. If the order is successful, the address will be displayed with a
blue marker. In case of out-of-stock, it will be displayed with a red 
marker.

Since the Faker class is used to generate the addresses, the cities 
do not actually exist and therefore cannot be displayed on the map. 
We introduced a new property in the connector configuration. It is 
called **onlineorders.cities** and contains a comma separated list 
of cities. For example:

onlineorders.cities=London,Paris,New York,Tokyo,Lima,Bangkok 

If this property does not exist, the connector will work as before. 
But, if it is specified, the *OnlineOrderGenerator* will randomly 
select one of these cities and assign it to the city in the shipping 
address when generating the order. 

Signed-off-by: Srecko Janjic <srecko.janjic@si.ibm.com>