fix: prevent historical events generated with future timestamps
Optionally, on startup for the first time, the Connector can
generate a week's worth of events with timestamps intentionally
backdated to make it look like the Connector has been running
and emitting events for a week. This is a useful feature for
rapidly setting up a demo suitable for using event stream
processing to perform historical analysis.

In very rare situations, this can produce events with a timestamp
that is greater than an hour after the current timestamp - which
results in an InvalidTimestampException.

This commit is a brute force protection against that by silently
discarding any events generated by DatagenHistoryGenerator that
have a future timestamp.

I did consider a more careful approach that identifies and fixes
exactly the circumstances that can result in a future
timestamp being generated, but that didn't feel warranted given
that this is only a synthetic data generator for demos. Plus this
approach will more easily catch any future regressions in this
area too.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>